refactor: Lotto의 패키지 변경 (#5)
